+++ shiny/render/transformer/__init__.py
@@ -4,14 +4,13 @@ from ._transformer import (  # noqa: F401
     OutputRenderer,
     output_transformer,
     is_async_callable,
-    resolve_value_fn,
     ValueFn,
+    ValueFnApp,  # pyright: ignore[reportUnusedImport]
     ValueFnSync,  # pyright: ignore[reportUnusedImport]
     ValueFnAsync,  # pyright: ignore[reportUnusedImport]
     TransformFn,  # pyright: ignore[reportUnusedImport]
     OutputTransformer,  # pyright: ignore[reportUnusedImport]
-    OutputRendererSync,  # pyright: ignore[reportUnusedImport]
-    OutputRendererAsync,  # pyright: ignore[reportUnusedImport]
+    resolve_value_fn,  # pyright: ignore[reportUnusedImport]
 )
 
 __all__ = (
@@ -21,5 +20,4 @@ __all__ = (
     "ValueFn",
     "output_transformer",
     "is_async_callable",
-    "resolve_value_fn",
 )